pyproject.toml
[project]
name = "exam-parser"
version = "0.1.0"
description = "试卷转 JSON 转换工具"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv>=1.0",
    "requests>=2.31",
    "img2pdf>=0.5",
    "jsonschema>=4.20",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
]

[project.scripts]
exam-parser = "exam_parser.main:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
